+++ internal/checks/labels_conflict_test.go
@@ -10,8 +10,12 @@ import (
 	"github.com/cloudflare/pint/internal/promapi"
 )
 
-func textExternalLabels(name, uri, k, v string) string {
+func textExternalLabelsRR(name, uri, k, v string) string {
 	return fmt.Sprintf("`%s` Prometheus server at %s external_labels already has %s=%q label set, please choose a different name for this label to avoid any conflicts.", name, uri, k, v)
+}
+
+func textExternalLabelsAR(name, uri, k, v string) string {
+	return fmt.Sprintf("This label is redundant. `%s` Prometheus server at %s external_labels already has %s=%q label set and it will be automatically added to all alerts, there's no need to set it manually.", name, uri, k, v)
 }
 
 func newLabelsConflict(prom *promapi.FailoverGroup) checks.RuleChecker {
@@ -35,8 +39,15 @@ func TestLabelsConflictCheck(t *testing.T) {
 			problems:    noProblems,
 		},
 		{
-			description: "no labels",
+			description: "no labels / recording",
 			content:     "- record: foo\n  expr: up == 0\n",
+			checker:     newLabelsConflict,
+			prometheus:  newSimpleProm,
+			problems:    noProblems,
+		},
+		{
+			description: "no labels / alerting",
+			content:     "- alert: foo\n  expr: up == 0\n",
 			checker:     newLabelsConflict,
 			prometheus:  newSimpleProm,
 			problems:    noProblems,
@@ -63,7 +74,7 @@ func TestLabelsConflictCheck(t *testing.T) {
 			},
 		},
 		{
-			description: "conflict",
+			description: "conflict / recording",
 			content:     "- record: foo\n  expr: up == 0\n  labels:\n    foo: bar\n",
 			checker:     newLabelsConflict,
 			prometheus:  newSimpleProm,
@@ -75,7 +86,7 @@ func TestLabelsConflictCheck(t *testing.T) {
 							Last:  4,
 						},
 						Reporter: checks.LabelsConflictCheckName,
-						Text:     textExternalLabels("prom", uri, "foo", "bob"),
+						Text:     textExternalLabelsRR("prom", uri, "foo", "bob"),
 						Details:  alertsExternalLabelsDetails("prom", uri),
 						Severity: checks.Warning,
 					},
@@ -89,8 +100,73 @@ func TestLabelsConflictCheck(t *testing.T) {
 			},
 		},
 		{
-			description: "no conflict",
+			description: "conflict / alerting / different",
+			content:     "- alert: foo\n  expr: up == 0\n  labels:\n    foo: bar\n",
+			checker:     newLabelsConflict,
+			prometheus:  newSimpleProm,
+			problems: func(uri string) []checks.Problem {
+				return []checks.Problem{
+					{
+						Lines: parser.LineRange{
+							First: 4,
+							Last:  4,
+						},
+						Reporter: checks.LabelsConflictCheckName,
+						Text:     textExternalLabelsRR("prom", uri, "foo", "bob"),
+						Details:  alertsExternalLabelsDetails("prom", uri),
+						Severity: checks.Warning,
+					},
+				}
+			},
+			mocks: []*prometheusMock{
+				{
+					conds: []requestCondition{requireConfigPath},
+					resp:  configResponse{yaml: "global:\n  external_labels:\n    foo: bob\n"},
+				},
+			},
+		},
+		{
+			description: "conflict / alerting / identical",
+			content:     "- alert: foo\n  expr: up == 0\n  labels:\n    foo: bar\n",
+			checker:     newLabelsConflict,
+			prometheus:  newSimpleProm,
+			problems: func(uri string) []checks.Problem {
+				return []checks.Problem{
+					{
+						Lines: parser.LineRange{
+							First: 4,
+							Last:  4,
+						},
+						Reporter: checks.LabelsConflictCheckName,
+						Text:     textExternalLabelsAR("prom", uri, "foo", "bar"),
+						Details:  alertsExternalLabelsDetails("prom", uri),
+						Severity: checks.Warning,
+					},
+				}
+			},
+			mocks: []*prometheusMock{
+				{
+					conds: []requestCondition{requireConfigPath},
+					resp:  configResponse{yaml: "global:\n  external_labels:\n    foo: bar\n"},
+				},
+			},
+		},
+		{
+			description: "no conflict / recording",
 			content:     "- record: foo\n  expr: up == 0\n  labels:\n    foo: bar\n",
+			checker:     newLabelsConflict,
+			prometheus:  newSimpleProm,
+			problems:    noProblems,
+			mocks: []*prometheusMock{
+				{
+					conds: []requestCondition{requireConfigPath},
+					resp:  configResponse{yaml: "global:\n  external_labels:\n    bob: bob\n"},
+				},
+			},
+		},
+		{
+			description: "no conflict / alerting",
+			content:     "- alert: foo\n  expr: up == 0\n  labels:\n    foo: bar\n",
 			checker:     newLabelsConflict,
 			prometheus:  newSimpleProm,
 			problems:    noProblems,